% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oulad.R
\docType{data}
\name{assessment}
\alias{assessment}
\title{Assessment data from OULAD dataset}
\format{A data frame with 206 rows and 6 variables:
\describe{
  \item{code_module}{Name of course, to which assessment belongs}
  \item{code_presentation}{Name of semester, to which assessment belongs}
  \item{id_assessment}{Unique integer value identifiing assessment}
  \item{assessment_type}{Identifies type of assessment:}
  \item{date}{Assessment cut-off date in days from official start of course semester}
  \item{weight}{Weight of assessment for the final grade from the course}
}}
\source{
\url{https://analyse.kmi.open.ac.uk/open_dataset}
}
\usage{
data(assessment)
}
\description{
A dataset containing infromation about assessments in each course-semester
included in OULAD.
}
\section{Assessment type values}{

\itemize{
  \item{TMA - Tutor Marked Assessment, usually essay  or report marked by human}
  \item{CMA - Computer Marked Assessment, usually some problem solved within Virtual Learning Environment and graded by program}
  \item{Exam - Final exam of the course semester}
}
}
\seealso{
\code{\link{oulad}}
}
\keyword{datasets}

