% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oulad.R
\docType{data}
\name{course}
\alias{course}
\title{Courses data from OULAD dataset}
\format{A data frame with 22 rows and 3 variables:
\describe{
  \item{code_module}{Anonymised name of the course (called module at Open University)}
  \item{code_presentation}{Name of course semester (called presentation), J is for winter and B for summer}
  \item{module_presentation_length}{Length of course in days from official start}
}}
\source{
\url{https://analyse.kmi.open.ac.uk/open_dataset}
}
\usage{
data(course)
}
\description{
A dataset containing names of courses accompanied with the semestral
information - name and length.
}
\keyword{datasets}

