% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loaders.R
\name{load_studentVle}
\alias{load_studentVle}
\title{Convenience: load a slice of studentVle}
\usage{
load_studentVle(module = NULL, presentation = NULL, date_between = NULL)
}
\arguments{
\item{module}{Optional code module (character scalar)}

\item{presentation}{Optional code presentation (character scalar)}

\item{date_between}{Optional numeric(2) for inclusive date window (OULAD uses day offsets)}
}
\description{
Convenience: load a slice of studentVle
}
