% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-loaders.R
\name{open_studentVle}
\alias{open_studentVle}
\title{Open the partitioned studentVle dataset (lazy)}
\usage{
open_studentVle()
}
\value{
An Arrow Dataset; use dplyr verbs then \code{collect()}
}
\description{
Open the partitioned studentVle dataset (lazy)
}
\examples{
# Read only one presentation
# open_studentVle() |>
#   dplyr::filter(code_module == "AAA", code_presentation == "2013B") |>
#   dplyr::collect()
}
