% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oulad.R
\docType{data}
\name{student_assessment}
\alias{student_assessment}
\title{Student assessments results data from OULAD daset}
\format{A data frame with 173912 rows and 5 variables:
\describe{
   \item{id_assessment}{Unique identifier of assessment, connects to dataset assessment}
   \item{id_student}{Unique identifier of student, connects to dataset student}
   \item{date_submitted}{Date when student submitted the assessment in days from official start of the course}
   \item{is_banked}{Student transfered result of assessment from previous course attempt, logical}
   \item{score}{Score achieved by student in assessment}
}}
\source{
\url{https://analyse.kmi.open.ac.uk/open_dataset}
}
\usage{
data(student_assessment)
}
\description{
Dataset contains assessment results of each student in each of his attended course.
}
\seealso{
\code{\link{oulad}}
}
\keyword{datasets}

