% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oulad.R
\docType{data}
\name{student_vle}
\alias{student_vle}
\title{Student Virtual Learning Environment (VLE) interaction data from OULAD dataset}
\format{A data frame with 10655280 rows and 6 variables:
\describe{
   \item{code_module}{Course name, to which "resource" belongs}
   \item{code_presentation}{Semester name, to which "resource" belongs}
   \item{id_student}{Unique student identifier, connects to student dataset}
   \item{id_site}{Unique "resource" identifier, connects to vle dataset}
   \item{date}{Date of interaction between student and "resource" in days from
   official course start}
   \item{sum_click}{Number of interactions between student and "resource" on selected date}
}}
\source{
\url{https://analyse.kmi.open.ac.uk/open_dataset}
}
\usage{
data(student_vle)
}
\description{
Dataset provides daily summary log of student interaction with "resources" in
VLE.
}
\seealso{
\code{\link{oulad}}
}
\keyword{datasets}

