% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oulad.R
\docType{data}
\name{vle}
\alias{vle}
\title{Virtual Learning Environment (VLE) struture data from OULAD dataset}
\format{A data frame with 6364 rows and 6 variables:
\describe{
   \item{id_site}{Unique "resource" identifier}
   \item{code_module}{Course, to which "resource" belongs}
   \item{code_presentation}{Semester name, to which "resource" belongs}
   \item{activity_type}{Type of resource}
   \item{week_from}{Start week from the official course start, in which
   "resource" is recommended for study, NA means not in study plan}
   \item{week_to}{End week from the official course start, in which student
   should finish with the materials presented in "resource", NA means not in study plan}
}}
\source{
\url{https://analyse.kmi.open.ac.uk/open_dataset}
}
\usage{
data(vle)
}
\description{
The dataset provides the information about VLE "resources" of each course presented in
OULAD dataset. By the "resources" we mean different types of web pages with the
course relevant information. Resources can be included into study plan for the course.
}
\seealso{
\code{\link{oulad}}
}
\keyword{datasets}

