% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oulad.R
\docType{data}
\name{student}
\alias{student}
\title{Student data from OULAD dataset}
\format{A data frame with 32593 rows and 12 variables:
\describe{
}}
\usage{
data(student)
}
\description{
A dataset containing demographical information about students, their courses visited
and final results each course.
}
\keyword{datasets}

